feature: allow empty string for indent char
1. Check if `char` variable is set and only insert indent character when
   not empty. Insert additional character from real underlying
   whitespace instead.
2. Break out conditions that are used multiple times into their own
   variables for easier maintenance.
3. Add documentation for char="".
4. Update README.md to reflect new option.

When setting the `char = ""` before, the resulting virtual text would
have been too short since it won't insert any character as
a replacement.  This also affected the highlighting of the virtual text.
Additionally when no indent char was desired, the only way to "ignore"
it was to set `char` to a space. This, however, does not look good when
also displaying `listchars` because it will insert an empty space even
when the user has set a different character to be displayed in place of
a space.