Add `CursorMovedI` event to `IndentBlanklineRefresh` autocmd
It is obvious that users shouldn't move around in insert mode, however
in some scenarios the cursor moves out of the current context in insert
mode.

For example let's say we are editing something like this
```c
if (myBool) {

} // <-- cursor is on this line
```
if the user presses `o` the cursor moves out of the if statement context
with the `CursorMovedI` event being called instead of `CursorMoved`.

Or if the user has an autopair plugin by pressing `{<cr>` that plugin
potentially creates a new context and moves the cursor inside that
context in insert mode.