fix: config overwrite
The documentation was not consistent with the code and comments.
I think the version that is described in the documentation is more
useful, so this should be what the code does.

I think it is fair to classify this as a bug, so this is not a breaking
change. But it does behave differently now. If you want the old
behavior, call `ibl.setup` once before.