feat: support empty foldtext in nvim 0.10
Neovim 0.10 changed the behavior for when `'foldtext'` is empty.
It now displays the line normally with highlighting.

`vim.fn.foldtextresult` does however not reflect this change and returns
the previous foldtext, like `+--  4 lines folded`. This breaks the
current implementation which checks if indentation lines can be
displayed over the foldtext.

So until (if ever) this gets fixed, we need to do a custom check.

fix #901